requirements.txt
tensorflow
opencv-python
pillow
numpy
